Updating to the newest version

+++ BudgedManager/Models/Entity/Expense.cs
@@ -3,9 +3,12 @@
 public class Expense
 {
     public int Id { get; set; }
-    public double Amount { get; set; }
+
+    public decimal Amount { get;set; }
     public int CategoryId { get; set; }
     public Category? Category { get; set; }
     public DateTime Date { get; set; }
     public string? Comment { get; set; }
+    
+    
 }
